# Performance improvements to generation of top accounts page