--- a/apps/block_scout_web/lib/block_scout_web/templates/address/_tile.html.eex
+++ b/apps/block_scout_web/lib/block_scout_web/templates/address/_tile.html.eex
@@ -15,7 +15,7 @@
           <!-- number of txns for this address -->
           <span class="mr-4 mb-2">
             <span data-test="transaction_count">
-              <%= transaction_count(@address) %>
+              <%= @tx_count %>
             </span> <%= gettext "Transactions sent" %>
             <% if validator?(@address) do %>
               <span data-test="validation_count">
@@ -36,7 +36,7 @@
               data-usd-exchange-rate="<%= @exchange_rate.usd_value %>">
             </span>
             <!-- percentage of coins from total supply -->
-            <span class="ml-0 ml-md-2">(<%= balance_percentage(@address) %>)</span>
+            <span class="ml-0 ml-md-2">(<%= balance_percentage(@address, @total_supply) %>)</span>
           </div>
         </div>
       </div>