This PR addresses #873.

The list_top_addresses query is now a left join of addresses and transactions tables,
returning the top 250 address and their respective transaction counts at the same time.

The call to transaction_count() which hits the database has been removed from the per account tile template,
and the value is now passed in from the above query.

balance_percentage/2() has been added, taking total_supply as an argument rather than querying the database.
The per account tile template now uses this version, with total_supply queried once and passed in.

It may be worth adding an index to the addresses table i.e.
create index addresses_fetched_coin_balance_hash_index on addresses (fetched_coin_balance desc nulls last, hash asc);

Note the list_top_address query makes use of a SQL fragment as the version of Ecto used by blockscout does not support coalesce.