@svenski123 2018-11-03
Here are some before/after timings - transactions table had about 660K rows, addresses table had about 8.5K rows.

**BEFORE**
```
18:04:39.692 application=plug request_id=2lhqulcql36ovdnsqk00q3c1 [info] GET /accounts
18:04:42.613 application=phoenix [info] JOIN "exchange_rate:new_rate" to BlockScoutWeb.ExchangeRateChannel
  Transport:  Phoenix.Transports.WebSocket (2.0.0)
  Serializer:  Phoenix.Transports.V2.WebSocketSerializer
  Parameters: %{}
18:04:42.613 application=phoenix [info] Replied exchange_rate:new_rate :ok
18:04:42.614 application=phoenix [info] JOIN "blocks:indexing" to BlockScoutWeb.BlockChannel
  Transport:  Phoenix.Transports.WebSocket (2.0.0)
  Serializer:  Phoenix.Transports.V2.WebSocketSerializer
  Parameters: %{}
18:04:42.614 application=phoenix [info] Replied blocks:indexing :ok
18:05:03.092 application=plug request_id=2lhquj7r9j1730n16s00ntc5 [info] Sent 200 in 60432ms
18:05:35.724 application=indexer [error] realtime indexer failed to fetch beneficiaries for block 5400837: "Error fetching block reward contract beneficiaries".  Block will be retried by catchup indexer.
18:05:39.996 application=plug request_id=2lhqulcql36ovdnsqk00q3c1 [info] Sent 200 in 60303ms
```

**AFTER**
```
18:08:36.070 application=plug request_id=2lhqv353rfkosrem6c00d3s6 [info] GET /accounts
18:08:38.137 application=plug request_id=2lhqv353rfkosrem6c00d3s6 [info] Sent 200 in 2066ms
18:08:39.424 application=plug request_id=2lhqv3bbpsv56neutk00his1 [info] GET /accounts
18:08:39.665 application=phoenix [info] JOIN "exchange_rate:new_rate" to BlockScoutWeb.ExchangeRateChannel
  Transport:  Phoenix.Transports.WebSocket (2.0.0)
  Serializer:  Phoenix.Transports.V2.WebSocketSerializer
  Parameters: %{}
18:08:39.665 application=phoenix [info] Replied exchange_rate:new_rate :ok
18:08:39.666 application=phoenix [info] JOIN "blocks:indexing" to BlockScoutWeb.BlockChannel
  Transport:  Phoenix.Transports.WebSocket (2.0.0)
  Serializer:  Phoenix.Transports.V2.WebSocketSerializer
  Parameters: %{}
18:08:39.666 application=phoenix [info] Replied blocks:indexing :ok
18:08:40.869 application=plug request_id=2lhqv3bbpsv56neutk00his1 [info] Sent 200 in 1444ms
```