@acravenho 2018-11-05
@feliperenan I think we should split this into 2 separate issues. While this PR may not work on Mainnet yet, we could do what we did with the token holders and split the active coin balance into another table which should make this much quicker. This is a big improvement on what is currently available. We could use this as is on 7 out of 8 chains and make the necessary changes in a separate PR to make this work on Mainnet.